refactor: split ContractRunnerPanel into stateful sub-components
